Move storage tests.
